* (Окружение) бочки со взрывчаткой * (Окружение) создать звуки взрывов бочек

+++ part_01_base_settings/Assets/Scripts/Gun/NewGunController.cs
@@ -25,7 +25,7 @@ public class NewGunController : MonoBehaviour
 
 
     public Transform firePoint; //информация из GunSettings
-    public float damage = 25f; //информация из GunSettings
+    public int damage = 25; //информация из GunSettings
 
 
     private IEnumerator shootBulletCoroutine;
